requirements.txt
accelerate==1.7.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.4
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
braceexpand==0.1.7
certifi==2025.4.26
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
datasets==3.6.0
decord==0.6.0
dill==0.3.8
docker-pycreds==0.4.0
einops==0.7.0
einx==0.1.3
filelock==3.18.0
fonttools==4.58.0
frozendict==2.4.6
frozenlist==1.6.0
fsspec==2025.3.0
gitdb==4.0.12
gitpython==3.1.44
huggingface-hub==0.29.1
idna==3.10
ipdb==0.13.13
jinja2==3.1.6
kiwisolver==1.4.8
lmdb==1.6.2
loguru==0.7.3
markupsafe==3.0.2
matplotlib==3.7.0
mediapy==1.1.6
mpmath==1.3.0
multidict==6.4.4
multiprocess==0.70.16
networkx==3.4.2
ninja==1.11.1.4
numpy==1.24.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.7.0.72
packaging==25.0
pandas==2.2.3
pillow==11.2.1
propcache==0.3.1
protobuf==6.31.0
psutil==7.0.0
pyarrow==20.0.0
pydantic==2.11.4
pydantic-core==2.33.2
pyparsing==3.2.3
pytz==2025.2
pyyaml==6.0.2
requests==2.32.3
safetensors==0.4.5
scipy==1.10.1
sentencepiece==0.1.99
sentry-sdk==2.29.1
setproctitle==1.3.6
smmap==5.0.2
sympy==1.13.1
tokenizers==0.21.1
tomli==2.2.1
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.49.0
triton==3.1.0
typing-extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wandb==0.19.11
webdataset==0.2.111
xxhash==3.5.0
yarl==1.20.0